requirements.txt
streamlit==1.41.1
opentelemetry-distro==0.50b0
opentelemetry-exporter-otlp=1.29.0